adjusted the button test to run on cirrus + week6 summary
