+++ app/src/androidTest/java/com/epfl/neighborfood/neighborfoodandroid/ui/activities/DisplayMessageActivityTest.java
@@ -1,4 +1,4 @@
-package com.epfl.neighborfood.neighborfoodandroid.ui.layouts;
+package com.epfl.neighborfood.neighborfoodandroid.ui.activities;
 
 
 import static androidx.test.espresso.Espresso.onView;
@@ -16,7 +16,6 @@ import androidx.test.espresso.matcher.ViewMatchers;
 import androidx.test.ext.junit.runners.AndroidJUnit4;
 
 import com.epfl.neighborfood.neighborfoodandroid.R;
-import com.epfl.neighborfood.neighborfoodandroid.ui.activities.DisplayMessageActivity;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
