+++ app/src/androidTest/java/com/epfl/neighborfood/neighborfoodandroid/ui/activities/ProfileEditingActivityTest.java
@@ -1,4 +1,4 @@
-package com.epfl.neighborfood.neighborfoodandroid.ui.layouts;
+package com.epfl.neighborfood.neighborfoodandroid.ui.activities;
 
 import static androidx.test.espresso.Espresso.onView;
 import static androidx.test.espresso.action.ViewActions.click;
@@ -8,7 +8,6 @@ import static androidx.test.espresso.intent.Intents.intended;
 import static androidx.test.espresso.intent.Intents.intending;
 import static androidx.test.espresso.intent.matcher.IntentMatchers.hasAction;
 import static androidx.test.espresso.intent.matcher.IntentMatchers.hasData;
-import static androidx.test.espresso.intent.matcher.IntentMatchers.toPackage;
 import static androidx.test.espresso.matcher.ViewMatchers.assertThat;
 import static androidx.test.espresso.matcher.ViewMatchers.withId;
 import static androidx.test.espresso.matcher.ViewMatchers.withText;
@@ -19,7 +18,6 @@ import static org.hamcrest.core.AllOf.allOf;
 import static org.junit.Assert.assertTrue;
 
 import android.app.Activity;
-import android.app.Application;
 import android.app.Instrumentation;
 import android.content.Intent;
 import android.content.res.Resources;
@@ -28,31 +26,21 @@ import android.os.Bundle;
 import android.provider.MediaStore;
 
 
-import androidx.lifecycle.Lifecycle;
 import androidx.test.core.app.ApplicationProvider;
 import androidx.test.espresso.intent.Intents;
 import androidx.test.ext.junit.rules.ActivityScenarioRule;
-import androidx.test.ext.junit.runners.AndroidJUnit4;
 import androidx.test.platform.app.InstrumentationRegistry;
 
-import com.epfl.neighborfood.neighborfoodandroid.AppContainerTestImplementation;
 import com.epfl.neighborfood.neighborfoodandroid.NeighborFoodApplication;
 import com.epfl.neighborfood.neighborfoodandroid.R;
 import com.epfl.neighborfood.neighborfoodandroid.models.User;
 import com.epfl.neighborfood.neighborfoodandroid.models.UserTestImplementation;
 import com.epfl.neighborfood.neighborfoodandroid.repositories.AuthRepository;
-import com.epfl.neighborfood.neighborfoodandroid.repositories.AuthRepositoryTestImplementation;
-import com.epfl.neighborfood.neighborfoodandroid.ui.activities.MainActivity;
-import com.epfl.neighborfood.neighborfoodandroid.ui.activities.ProfileEditingActivity;
 
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
-import org.junit.rules.RuleChain;
-import org.junit.runner.RunWith;
-
-import javax.inject.Inject;
 
 
 //@RunWith(AndroidJUnit4.class)
@@ -75,7 +63,7 @@ public class ProfileEditingActivityTest {
     @Test
     public void buttonSaveTest(){
 
-        onView(withId(R.id.saveButton)).perform(click());
+        onView(withId(R.id.saveButton)).perform(scrollTo(),click());
         //assertTrue(testRule.getScenario().getState() == Lifecycle.State.DESTROYED);
 
     }
