+++ app/src/main/res/layout/activity_profile_editing.xml
@@ -20,7 +20,7 @@
     <ScrollView
         android:id="@+id/profileEditScrollView"
         android:layout_width="match_parent"
-        android:layout_height="wrap_content">
+        android:layout_height="match_parent">
 
         <RelativeLayout
             android:id="@+id/relativeLayout"
@@ -103,6 +103,7 @@
                 android:layout_toEndOf="@id/profilePictureImageView"
                 android:text="foulen99"
                 android:textColor="#757575" />
+
             <TextView
                 android:id="@+id/emailLabel"
                 android:layout_width="match_parent"
@@ -147,16 +148,16 @@
                 android:hint="@string/bio"
                 android:minHeight="150dp"
                 tools:ignore="DuplicateSpeakableTextCheck" />
+            <Button
+                android:id="@+id/saveButton"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:layout_below="@id/bioValue"
+                android:onClick="onClick"
+                android:text="@string/save" />
+
         </RelativeLayout>
     </ScrollView>
-
-    <Button
-        android:id="@+id/saveButton"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_gravity="end|bottom"
-        android:onClick="onClick"
-        android:text="@string/save" />
 
 
 </LinearLayout>
