+++ summaries/week6.md
@@ -1,0 +1,18 @@
+# Summary for sprint 6
+
+## Badis (scrum master)
+
+## Elior
+
+## Khalil
+I had quite some trouble having the tests work with the chosen architecture, so I decided to go with a different approach. 
+At least now, the tests are now up and running, the feature of displaying the connected user works well, displaying the basic activities.
+My time estimates are still a bit off, but they are getting better.
+
+## Raed
+
+## Yassine 
+
+
+## Overall team
+
